Update dockerfile to use non root user
Running as root is annoying because any file created
while inside the container will be owned by root and need to
be cleaned up with sudo

This required hardcoding the uid to 1000 to match my uid outside
the container. This may not work on systems that have the primary user
id as something other than 1000 (1001 for instance).